Fix NPM Cache directory permission and PID Signalling bugs
NPM appears to not propagate the SIGTERM signal to the child node app
on alpine. This is fixed by unwrapping the call and starting the app
directly instead. NPM cache directory permission issues have appeared
again for some external reason, so we need to mitigate by explicitly
forcing a permission update on the cache directory to be owned by a
non-root user.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>