Ensure config disabled helm deployments still pull in appropriate secrets
There was a bug where the only way to ensure keycloak authentication was
enabled was to have config.configMap.KC_ENABLED defined, even when we are
not wanting to generate our own release-scoped config map. This fix adds in
a feature toggle to declare authentication mode intent independent of the
configuration.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>